Move the neopixel update to a periodic task
This also changes up the pixel animation to a nice marching wheel animation.
In my experiments, the PyPortal isn't really able to update the neopixels
any faster than around 10hz due to the time it takes to draw items on
the screen. This does, however, seem to make touch inputs much more
responsive.